requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cssselect2==0.5.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.3
django-cleanup==6.0.0
django-extensions==3.1.5
django-htmx==1.9.0
django-import-export==2.7.1
django-js-asset==2.0.0
django-markdownx==3.0.1
django-mptt==0.13.4
django-mptt-admin==2.2.0
django-solo==2.0.0
django-tinymce==3.4.0
et-xmlfile==1.1.0
fonttools==4.30.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.2
kombu==5.2.4
Markdown==3.3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.7
pyphen==0.12.0
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
sentry-sdk==1.5.7
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tinycss2==1.1.1
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.2
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
zopfli==0.2.1
